Build 0.3.0

+++ components/EntriesTable.tsx
@@ -28,6 +28,11 @@ export default function EntriesTable({
   const [filterText, setFilterText] = useState('');
 
   const handleDelete = async (entryId: string) => {
+    const confirmed = window.confirm(
+      'Are you sure you want to delete this entry?'
+    );
+    if (!confirmed) return;
+
     console.log(`Delete entry with ID: ${entryId}`);
 
     const { error } = await deleteEntry(entryId); // also gets {message}
@@ -123,16 +128,16 @@ export default function EntriesTable({
           (user?.role === 'admin' ||
             user?.role === 'superadmin' ||
             ownerClerkId === user?.clerkUserId);
-        console.log(
-          'Row User:',
-          ownerClerkId,
-          'Current User:',
-          user?.clerkUserId,
-          'Can edit:',
-          canEdit,
-          'for project:',
-          row.itemTitle
-        );
+        // console.log(
+        //   'Row User:',
+        //   ownerClerkId,
+        //   'Current User:',
+        //   user?.clerkUserId,
+        //   'Can edit:',
+        //   canEdit,
+        //   'for project:',
+        //   row.itemTitle
+        // );
         const LinkOutUrl = row.url ?? undefined;
         const LinkOut = LinkOutUrl ? (
           <Link href={LinkOutUrl} target="_blank">
